requirements.txt
# Dependencies
# Usage:
#   $ pip install -r requirements.txt

# Core Capirca dependencies
absl-py
ply
PyYAML
six>=1.12.0
typing_extensions

# Phase 2 API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy>=2.0
alembic>=1.11.0
requests>=2.31.0
